+++ app/src/main/java/woowacourse/shopping/ui/theme/Type.kt
@@ -9,13 +9,13 @@ import androidx.compose.ui.unit.sp
 val Typography =
     Typography(
         bodyLarge =
-            TextStyle(
-                fontFamily = FontFamily.Default,
-                fontWeight = FontWeight.Normal,
-                fontSize = 16.sp,
-                lineHeight = 24.sp,
-                letterSpacing = 0.5.sp,
-            ),
+        TextStyle(
+            fontFamily = FontFamily.Default,
+            fontWeight = FontWeight.Normal,
+            fontSize = 16.sp,
+            lineHeight = 24.sp,
+            letterSpacing = 0.5.sp,
+        ),
         /* Other default text styles to override
          * titleLarge = TextStyle(
          *     fontFamily = FontFamily.Default,
